Integrate collaborative crosswords
I added an API endpoint to the crossword app (which is currently running on heroku).
This solving software uses the API to make crosswords (similar to how it makes googles sheets).

The URLs to the crossword sheets are stored in a new table.
Users can manually add crossword sheets (as many per puzzle
as desired) with a button in the puzzle info panel.

The crossword sheets go in the same frame as the the google sheet. The user can tab the google
sheet and the crossword sheet(s). Also, the view defaults to the crossword sheet when there is
at least one. I think that's a reasonable assumption, although it could potentially be annoying,
since there is no way to delete a crossword sheet once it is made.

There is also a button to delete crossword sheets.